The 20th tutorial in the 30 Vanilla JS tutorials: a lesson in working with the SpeechRecognition API
The 20th tutorial in the 30 Vanilla JS tutorials: a lesson in working with the SpeechRecognition API; "The SpeechRecognition interface of the Web Speech API is the controller interface for the recognition service; this also handles the SpeechRecognitionEvent sent from the recognition service". On implementing the change in Chrome I've used 'webkitSpeechRecognition' and for Firefox this was 'SpeechRecognition'.